+++ ADM_Scada.Core/Models/ProductModel1.cs
@@ -10,14 +10,14 @@ namespace ADM_Scada.Core.Models
         public string ProdFullName { get; set; }
         public string HashCode { get; set; }
         public string Ingredient { get; set; }
-        public float? Exp { get; set; }
+        public decimal? Exp { get; set; }
         public string Market { get; set; }
         public string ProdName { get; set; }
         public string LabelPath { get; set; }
         public string Barcode { get; set; }
         public string DelayM4 { get; set; }
         public string DelayM5 { get; set; }
-        public float PackSize { get; set; }
+        public decimal PackSize { get; set; }
         public string LooseUom { get; set; }
         public string WholeUom { get; set; }
         public DateTime? CreatedDate { get; set; }
